Move attributes deserializers to one place
